+++ tools/build_inp_blocks.py
